src/aoc_fruit_detector/scripts/requirements.txt
absl-py==1.4.0
aggdraw==1.3.16
alabaster==0.7.16
antlr4-python3-runtime==4.9.3
apturl==0.5.2
astunparse==1.6.3
babel==2.16.0
bcrypt==3.2.0
black==24.4.2
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
cloudpickle==3.0.0
cmake==3.26.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==3.0.11
dbus-python==1.2.18
defer==1.0.6
-e git+https://github.com/facebookresearch/detectron2.git@2a420edb307c9bdf640f036d3b196bed474b8593#egg=detectron2
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.20.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.10.0
flatbuffers==23.1.21
fonttools==4.38.0
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.3
h5py==3.8.0
hdf5storage==0.1.19
httplib2==0.20.2
hydra-core==1.3.2
idna==3.4
imageio==2.34.2
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.9
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
keras-nlp==0.4.0
keyring==23.5.0
kiwisolver==1.4.4
kmeans-pytorch==0.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==15.0.6.1
lit==16.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.2
mat73==0.65
matplotlib==3.7.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.0
nibabel==5.0.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
paramiko==2.9.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.4.0
platformdirs==4.2.2
portalocker==2.10.1
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycocotools==2.0.8
pycups==2.0.1
pydot==1.4.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
pysptools==0.15.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-version==0.0.2
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2024.7.24
tk==0.1.0
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.4
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
visualkeras==0.0.2
wadllib==1.3.6
Werkzeug==2.2.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
yacs==0.1.8
zipp==3.14.0
